requirements.txt
streamlit
rdkit-pypi
pillow